# perf: use `ValueListBuilder` for `SeparatedSyntaxList`