Add `ValueListBuilder` to `SeparatedSyntaxList` saving 0.9 MB.

I compared the length of the input `SeparatedSyntaxList` and returned `doc` `List<Doc>`, (lengths are left and right respectively). 70% of all output lists are under 8 values, so I chose a default capacity of 8 for the scratch buffer.

`SeparatedSyntaxList.Print` will add at most 3 items to the output list per item in `SeparatedSyntaxList<T> list`, knowing this I decided that `ValueListBuilder` should rent an array from the `ArrayPool` that can contain all the values. This way we don't have to bother with the overhead of resizing.

![image](https://github.com/user-attachments/assets/b4ff9de1-8073-48a3-9c4a-a23240f58202)


### Benchmarks
#### Before
| Method                | Mean     | Error   | StdDev  | Gen0      | Gen1      | Gen2     | Allocated |
|---------------------- |---------:|--------:|--------:|----------:|----------:|---------:|----------:|
| Default_CodeFormatter | 105.5 ms | 0.93 ms | 0.73 ms | 3500.0000 | 2000.0000 | 500.0000 |  34.14 MB |
#### After
| Method                | Mean     | Error   | StdDev  | Gen0      | Gen1      | Allocated |
|---------------------- |---------:|--------:|--------:|----------:|----------:|----------:|
| Default_CodeFormatter | 104.6 ms | 2.07 ms | 4.76 ms | 3000.0000 | 1000.0000 |  33.25 MB |